refactor: extract reducer to a separate file to improve readability
